Third wave of Surface x64 fixes: agree on surfaceID,x,y,w,h,pitch,isMSB types
They were declared either long or int depending on files.
This is not 64bits compatible.
Since all those values will fit into an int, use int.

+++ platforms/Cross/plugins/SqueakFFIPrims/sqFFI.h
@@ -157,9 +157,9 @@ int ffiLogCallOfLength(void *nameIndex, int nameLength);
 void initSurfacePluginFunctionPointers();
 void initManualSurfaceFunctionPointers
 	(fn_ioRegisterSurface, fn_ioUnregisterSurface, fn_ioFindSurface);
-long createManualSurface
-	(long width, long height, long rowPitch, long depth, long isMSB);
-long destroyManualSurface(long surfaceID);
-long setManualSurfacePointer(long surfaceID, void* ptr);
+int createManualSurface
+	(int width, int height, int rowPitch, int depth, int isMSB);
+int destroyManualSurface(int surfaceID);
+int setManualSurfacePointer(int surfaceID, void* ptr);
 
 #endif /* SQ_FFI_H */
